Update package image and zoom functionality
This commit updates the package image in index.html and modifies the zoom functionality in index.css. The "SPECIAL" text has been changed to "FAMILY" in the HTML file. Additionally, the max-height property has been removed and the cursor style has been added to the package-zoom class in the CSS file.